+++ CHANGES.rst
@@ -41,6 +41,11 @@ ramp_fitting
 
 - Changed image units from e/s to DN/s (and added support for MJy/sr). Added gain reduction to convert to these units. [#1128]
 
+flux
+----
+
+- Create FluxStep to apply the flux correction to Level 2 data. [#1120]
+
 
 0.14.0 (2024-02-12)
 ===================
@@ -82,6 +87,8 @@ ramp_fitting
 ------------
 
 - Add default WCS when constructing image model from ramp model [#1072]
+
+- Account for Poisson noise from dark current when fitting ramps. [#1088]
 
 resample
 --------